refactor(tui): extract effectiveId into a shared selection module
Sidebar imported effectiveId from App, coupling the two UI entry points.
It's a pure selection helper — moved to ui/selection.ts so both import it
from a neutral module. Flagged by review.
